add custom hook for jotai
