+++ components/Light/index.js
@@ -1,20 +1,8 @@
-import { useState } from "react";
 import { LightButton, Icon, Text, Name, State } from "./Light.styled";
-import { useAtom } from "jotai";
-import { currentLights } from "../../store/lights";
+import { useLights } from "../../hooks/useLights";
 
-export default function Light({ name }) {
-  const [lights, setLights] = useAtom(currentLights);
-
-  const { id, isOn } = lights.find((light) => light.name === name);
-
-  function handleToggle(id) {
-    setLights(
-      lights.map((light) => {
-        return light.id === id ? { ...light, isOn: !light.isOn } : light;
-      })
-    );
-  }
+export default function Light({ name, id, isOn }) {
+  const { handleToggle } = useLights();
 
   return (
     <LightButton
